Add files via upload
Figured out how to promote widgets to add needed functionality.
