+++ centralwidgetcustom.h
@@ -1,0 +1,19 @@
+#ifndef CENTRALWIDGETCUSTOM_H
+#define CENTRALWIDGETCUSTOM_H
+
+#include <QWidget>
+
+class centralWidgetCustom : public QWidget
+{
+    Q_OBJECT
+public:
+    explicit centralWidgetCustom(QWidget *parent = nullptr);
+
+    ~centralWidgetCustom();
+
+signals:
+
+public slots:
+};
+
+#endif // CENTRALWIDGETCUSTOM_H
