+++ weaponslot.cpp
@@ -1,6 +1,21 @@
 #include "weaponslot.h"
 #include <QMessageBox>
 
+// constructor definition
+WeaponSlot::WeaponSlot(std::string fileName, long filePosition, short length)
+    : m_fileName(fileName), m_filePosition(filePosition), m_length(length)
+{
+    //m_file.open("C:\\Program Files (x86)\\Konami\\The Regiment\\system\\SASChars.u");
+    m_file.open(fileName);
+
+    // if file fails to open print error
+    if (!m_file)
+    {
+        QMessageBox msgBox(QMessageBox::Critical, "Error", "Error opening SASChars.u  Please confirm this executable is in the same directory as the file SASChars.u.");
+        msgBox.exec();
+    }
+}
+
 // takes a gun slot and fills it with a specific gun
 void WeaponSlot::assignGun (const std::string &weapon)
 {
